build(deps): switch from structopt to clap v3
As clap v3 is now out, and the structopt features are integrated into
(almost as-is), structopt is now in maintenance mode: no new feature
will be added [[1]]. Consequently, it seems appropriate to switch to
clap v3 to benefit from the continued development.

clap v3 is slightly larger than structopt. This is the affect on some
basic metrics:

    | Metric                | Before | After |
    |-----------------------|--------|-------|
    | compile release (sec) | 6.25   | 8.00  |
    | binary size (MB)      | 6.2    | 6.3   |

Also note that we break away from the default clap v3 behavior in one
case: exit codes. In the transition for v2 to v3, clap switched from an
exit status of 1 to an exit status of 2. For compatibility with previous
version of `choose`, we still return an exit code of 1, rather than 2.

[1]: https://docs.rs/structopt/latest/structopt/#maintenance

Tested:
    Ran `cargo test`:
        test result: ok. 206 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.02s

    Ran `test/e2e_test.sh`:
        All tests passed